tweak fields in redesign
