+++ app/javascript/mastodon/features/account_timeline/common.ts
@@ -1,0 +1,5 @@
+import { isClientFeatureEnabled } from '@/mastodon/utils/environment';
+
+export function isRedesignEnabled() {
+  return isClientFeatureEnabled('profile_redesign');
+}
