+++ app/javascript/mastodon/features/account_timeline/components/redesign.module.scss
@@ -1,0 +1,5 @@
+.number_field_wrapper {
+  a {
+    font-weight: unset;
+  }
+}
